refactor(graphql): restructure schema into feature-based modules
