+++ src/graphql/modules/job/resolvers.ts
@@ -1,7 +1,7 @@
 export const jobResolvers = {
     Query: {
         jobs: async (_: any, __: any, { prisma }: any) => {
-            return prisma.job.findMany()
-        }
-    }
-}
+            return prisma.job.findMany();
+        },
+    },
+};
